+++ exonum-java-binding-common/src/main/java/com/exonum/binding/common/proofs/map/MapEntry.java
@@ -16,15 +16,18 @@
 
 package com.exonum.binding.common.proofs.map;
 
-import java.util.Arrays;
+import static com.google.common.base.Preconditions.checkNotNull;
+
+import com.google.protobuf.ByteString;
+import java.util.Objects;
 
 /**
  * A map entry: a key-value pair. This entry does not permit null keys and values.
  */
-public class MapEntry {
-  private final byte[] key;
+public final class MapEntry {
+  private final ByteString key;
 
-  private final byte[] value;
+  private final ByteString value;
 
   /**
    * Creates a new entry in a flat map proof corresponding to a leaf node.
@@ -32,17 +35,27 @@ public class MapEntry {
    * @param value a value mapped to the key
    */
   public MapEntry(byte[] key, byte[] value) {
-    this.key = key;
-    this.value = value;
+    this.key = ByteString.copyFrom(key);
+    this.value = ByteString.copyFrom(value);
+  }
+
+  /**
+   * Creates a new entry in a flat map proof corresponding to a leaf node.
+   * @param key a node key
+   * @param value a value mapped to the key
+   */
+  public MapEntry(ByteString key, ByteString value) {
+    this.key = checkNotNull(key, "key");
+    this.value = checkNotNull(value, "value");
   }
 
   /** Returns the key in this entry. */
-  public byte[] getKey() {
+  public ByteString getKey() {
     return key;
   }
 
   /** Returns the value in this entry. */
-  public byte[] getValue() {
+  public ByteString getValue() {
     return value;
   }
 
@@ -55,11 +68,11 @@ public class MapEntry {
       return false;
     }
     MapEntry that = (MapEntry) o;
-    return Arrays.equals(key, that.key) && Arrays.equals(value, that.value);
+    return Objects.equals(key, that.key) && Objects.equals(value, that.value);
   }
 
   @Override
   public int hashCode() {
-    return Arrays.hashCode(key) ^ Arrays.hashCode(value);
+    return Objects.hash(key, value);
   }
 }