+++ exonum-java-binding-core/src/main/java/com/exonum/binding/storage/indices/ProofListIndexProxy.java
@@ -22,7 +22,9 @@ import static com.exonum.binding.storage.indices.StoragePreconditions.checkIndex
 import static com.exonum.binding.storage.indices.StoragePreconditions.checkPositionIndex;
 
 import com.exonum.binding.common.hash.HashCode;
-import com.exonum.binding.common.proofs.list.ListProof;
+import com.exonum.binding.common.proofs.list.ListProofNode;
+import com.exonum.binding.common.proofs.list.UncheckedListProof;
+import com.exonum.binding.common.proofs.list.UncheckedListProofAdapter;
 import com.exonum.binding.common.serialization.CheckingSerializerDecorator;
 import com.exonum.binding.common.serialization.Serializer;
 import com.exonum.binding.common.serialization.StandardSerializers;
@@ -155,12 +157,14 @@ public final class ProofListIndexProxy<E> extends AbstractListIndexProxy<E>
    * @throws IndexOutOfBoundsException if the index is invalid
    * @throws IllegalStateException if this list is not valid
    */
-  public ListProof getProof(long index) {
+  public UncheckedListProof getProof(long index) {
     checkElementIndex(index, size());
-    return nativeGetProof(getNativeHandle(), index);
+
+    ListProofNode listProofNode = nativeGetProof(getNativeHandle(), index);
+    return new UncheckedListProofAdapter<>(listProofNode, this.serializer);
   }
 
-  private native ListProof nativeGetProof(long nativeHandle, long index);
+  private native ListProofNode nativeGetProof(long nativeHandle, long index);
 
   /**
    * Returns a proof that some elements exist in the specified range in this list.
@@ -170,13 +174,16 @@ public final class ProofListIndexProxy<E> extends AbstractListIndexProxy<E>
    * @throws IndexOutOfBoundsException if the range is not valid
    * @throws IllegalStateException if this list is not valid
    */
-  public ListProof getRangeProof(long from, long to) {
+  public UncheckedListProof getRangeProof(long from, long to) {
     long size = size();
-    return nativeGetRangeProof(getNativeHandle(), checkElementIndex(from, size),
+    ListProofNode listProofNode = nativeGetRangeProof(getNativeHandle(),
+        checkElementIndex(from, size),
         checkPositionIndex(to, size));
+
+    return new UncheckedListProofAdapter<>(listProofNode, this.serializer);
   }
 
-  private native ListProof nativeGetRangeProof(long nativeHandle, long from, long to);
+  private native ListProofNode nativeGetRangeProof(long nativeHandle, long from, long to);
 
   /**
    * Returns the root hash of the proof list.